% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_log_tcga_files.R
\name{create_log_tcga_files}
\alias{create_log_tcga_files}
\title{create log tcga files}
\usage{
create_log_tcga_files()
}
\description{
Takes upper norm tcga files creates processed form with coding genes and cancer samples only. Goes through all cancers that have upper_norm normalized RNAseq stored on data2 
then removes normals, log2 x+1 transforms, and restricts to coding genes. Then prints out a file for each.
}
