% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_log_tcga_files_normal.R
\name{create_log_tcga_files_normal}
\alias{create_log_tcga_files_normal}
\title{create log tcga files}
\usage{
create_log_tcga_files_normal()
}
\description{
Takes upper norm tcga files creates processed form with coding genes and normal samples only. Goes through all cancers that have upper_norm normalized RNAseq stored on data2 and selects normal files 
then removes cancers, log2 x+1 transforms, and restricts to coding genes. Then prints out a file for each.
}
