% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.survival.R
\name{do.survival}
\alias{do.survival}
\title{does survival analysis}
\usage{
do.survival(threshold = 0, pattern, path, output_folder,
  component = "comp.1", pred_file = F)
}
\arguments{
\item{threshold}{threshold to use to pick samples}

\item{pattern}{pattern of prediction files}

\item{path}{location of prediction files}

\item{output_folder}{the output folder}

\item{component}{component to use}

\item{pred_file}{using scores or a file with predictions}
}
\description{
Reads file with features(PC's) in columns and samples in rows. And an annotation file.You pick a max # of features. Using caret
runs a prediction iteratively from 2 to max.features. Picks best predictor with least featurs. Returns samples correctly predicted
leaves out those samples incorrectly predicted.
}
