% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_full_boxplot_meth.R
\name{do_full_boxplot_meth}
\alias{do_full_boxplot_meth}
\title{create tcga boxplots}
\usage{
do_full_boxplot_meth(file_pattern, tcga_location, iorio.adeno_file,
  iorio.sclc_file, sclc_file, output_folder)
}
\arguments{
\item{tcga_location}{location of predicted files of tcga}

\item{iorio.adeno_file}{location of predicted file of iorio.adeno}

\item{iorio.sclc_file}{location of predicted file of iorio.sclc}

\item{sclc_file}{location of predicted file of sclc}

\item{output_folder}{the folder to output to, default is ./ i.e. current folder}

\item{pattern}{pattern to grep for in files that have predictions}
}
\description{
Creates boxplots from prediction scores from pls or plsda
Makes some boxplots
}
