% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_full_boxplot.R, R/do_full_boxplot_RNA.R
\name{do_full_boxplot_RNA}
\alias{do_full_boxplot_RNA}
\title{create tcga boxplots}
\usage{
do_full_boxplot_RNA(file_pattern, tcga_location, crpc_file, nepc_file,
  sclc_file, output_folder, comp.x = "comp.1", components = 3,
  threshold = 3)

do_full_boxplot_RNA(file_pattern, tcga_location, crpc_file, nepc_file,
  sclc_file, output_folder, comp.x = "comp.1", components = 3,
  threshold = 3)
}
\arguments{
\item{tcga_location}{location of predicted files of tcga}

\item{crpc_file}{location of predicted file of crpc}

\item{nepc_file}{location of predicted file of nepc}

\item{sclc_file}{location of predicted file of sclc}

\item{output_folder}{the folder to output to, default is ./ i.e. current folder}

\item{pattern}{pattern to grep for in files that have predictions}

\item{pattern}{pattern to grep for in files that have predictions}

\item{tcga_location}{location of predicted files of tcga}

\item{crpc_file}{location of predicted file of crpc}

\item{nepc_file}{location of predicted file of nepc}

\item{sclc_file}{location of predicted file of sclc}

\item{output_folder}{the folder to output to, default is ./ i.e. current folder}
}
\description{
Creates boxplots from prediction scores from pls or plsda
Makes some boxplots

Creates boxplots from prediction scores from pls or plsda
Makes some boxplots
}
