% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick.samples.R
\name{pick.samples}
\alias{pick.samples}
\title{Pick samples}
\usage{
pick.samples(data.file, max.features = 5, anno.file, train.method = "LOOCV",
  technique = "lda")
}
\arguments{
\item{data.file}{Filepath/filename of data matrix}

\item{max.features}{the maximum number of features to use}

\item{anno.file}{annotation file, has headers, first column is sample names, 2nd is type}

\item{train.method}{the method used for training, default is leave one out cross validation, check caret for options}

\item{technique}{the method used for model building, default is linear discriminant analysis, check caret for options (e.g. svmLinear,}
}
\description{
Reads file with features(PC's) in columns and samples in rows. And an annotation file.You pick a max # of features. Using caret
runs a prediction iteratively from 2 to max.features. Picks best predictor with least featurs. Returns samples correctly predicted
leaves out those samples incorrectly predicted.
}
