% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pls.R
\name{plot_pls}
\alias{plot_pls}
\title{Plot PLS}
\usage{
plot_pls(file, info.name, info.type, title = "", labels = TRUE,
  PCx = "comp.1", PCy = "comp.2", ellipse = F, conf = 0.95,
  saveplot = F, savename = "default", savetype = ".pdf", w = 8,
  h = 6, legendname = "type", colpalette = NULL)
}
\arguments{
\item{file}{File containing scores matrix}

\item{info.name}{Vector of sample names}

\item{info.type}{Vector of sample types in the same order}

\item{title}{Title of the plot}

\item{labels}{default=T}

\item{PCx, PCy}{PCs to display}

\item{ellipse}{Construct confidence region based on groups in info.type, default = T}

\item{conf}{default = 0.95}

\item{saveplot}{do you want to save the plot}

\item{savetype}{default=".pdf", ".png" is also possible}

\item{w}{width default=8}

\item{h}{height default=6}

\item{h}{colpalette can add your own vector of colors to color groups. default is ggplot default colors}

\item{h}{legendname legend name}
}
\description{
Plots PLS from scores file (output of PLSR_from_file)
}
