% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSR_from_file_and_predict_second_dataset.R
\name{PLSR_from_file_and_predict_second_dataset}
\alias{PLSR_from_file_and_predict_second_dataset}
\title{Partial Least Squares and predict second dataset}
\usage{
PLSR_from_file_and_predict_second_dataset(file, file2, sample.names,
  sample.type, y.response, sample.names2 = NULL, sample.type2 = NULL,
  train_string, test_string, title = "PLSR", comp.x = "comp.1",
  comp.y = "comp.2", comps = 2, labels = F, saveplot = T,
  savetype = ".png", w = 8, h = 6, legendname = "default", scale = F,
  plot_both = T, colpalette = NULL, shape.palette = NULL, ellipses = T,
  conf = 0.9, varimax = F, varimax.comp = 2, output_folder = "./",
  TCGA = F, threshold = 3)
}
\arguments{
\item{file}{file for X matirx}

\item{file2}{file for test data matrix}

\item{sample.names}{Vector of sample names in X matrix}

\item{sample.type}{vector of sample groups}

\item{y.response}{numeric vector of response values in same order as samples}

\item{sample.names2}{Vector of sample names in 2nd dataset, if needed}

\item{sample.type2}{Vector of sample types in 2nd dataset, if needed}

\item{train_string}{string of training data to insert in file name of predicted scores}

\item{test_string}{string of data being tested to insert in file name of predicted scores}

\item{title}{title of the plot}

\item{comp.x, comp.y}{comps to display}

\item{labels}{label the plot, default = T}

\item{saveplot}{whether to save the plot, default is F}

\item{savetype}{the type of plot to save,options are ".pdf" or ".png"}

\item{w}{is width of plot to be saved}

\item{h}{is height of plot to be saved}

\item{legendname}{is the legend name}

\item{scale}{default=T}

\item{plot_both}{if true plots both training and test set in color}

\item{colpalette}{allows you to put in a color palette of form c("#F87660", "#39B600",....etc) to manually assign colors}

\item{shape.palette}{allows you to put in a shape palette of form c(1, 3,....etc) to manually assign shapes}

\item{varimax}{If T performs Varimax rotation,}

\item{varimax.comp}{# of varimax components, kind of hacky, keep this # the same as # of comps. Will fix later.}

\item{output_folder}{the folder to output to, default is ./ i.e. current folder}

\item{TCGA}{predicted files are from TCGA, barcodes separated by periods, so remove normal samples, default is FALSE}

\item{comp}{number of components to compute}
}
\description{
Builds PLS model from training dataset and predicts second dataset
Writes out predicted.scores of the second dataset
Plots original PLS, projected samples only, and projected samples ontop of original PLS
}
