% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_PCA_graph.R
\name{predict_PCA_graph}
\alias{predict_PCA_graph}
\title{Predict all files in a folder using PCA of a training dataset, file names have to be of form name_.......txt}
\usage{
predict_PCA_graph(prediction_file, test_files_folder, anno.file,
  output_folder = "./", train_string = "", test_pattern = "", comps = 2,
  rotate = F, varimax = F, varimax.comp = 2, shape.palette = NULL)
}
\arguments{
\item{prediction_file}{file  with training data features rows , genes columns, features are in column 1}

\item{test_files_folder}{folder of test files}

\item{anno.file}{Annotation file, first column is sample names, 2nd is annotation}

\item{output_folder}{output folder}

\item{train_string}{string of training data to insert in file name of predicted scores}

\item{comps}{number of components to compute}

\item{rotate}{do you want to rotate +/- signs, default=F.}

\item{varimax.comp}{# of varimax comps , has to match # of comps}

\item{shape.palette}{vector with numbers reprsenting shapes according to R plotting}

\item{test_string}{string of data being tested to insert in file name of predicted scores}

\item{varimax.it}{do you want to perform varimax , default varimax=F}
}
\description{
Predict all files in a folder using PCA of a training dataset, file names have to be of form name_.......txt
}
