% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_PLSDA_graph.R
\name{predict_PLSDA_graph}
\alias{predict_PLSDA_graph}
\title{Predict all files in a folder using PLSDA of a training dataset, file names have to be of form name_.......txt
Predicts classifications of test dataset and writes out predictions}
\usage{
predict_PLSDA_graph(prediction_file, test_files_folder, anno.file,
  output_folder = "./", train_string = "", test_pattern = "",
  train_pattern = "", comps = 3)
}
\arguments{
\item{prediction_file}{file  with training data features rows , genes columns, features are in column 1}

\item{test_files_folder}{folder of test files}

\item{anno.file}{Annotation file, first column is sample names, 2nd is annotation}

\item{output_folder}{output folder}

\item{train_string}{string of training data to insert in file name of predicted scores}

\item{train_pattern}{annotation type of main group in your comparison ,taken from 2nd column in anno.}

\item{comps}{number of components to compute}

\item{test_string}{string of data being tested to insert in file name of predicted scores}
}
\description{
Predict all files in a folder using PLSDA of a training dataset, file names have to be of form name_.......txt
Predicts classifications of test dataset and writes out predictions
}
