% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_to_gene.R
\name{seg_to_gene}
\alias{seg_to_gene}
\title{does survival analysis}
\usage{
seg_to_gene(seg_file, gene_file, output_name)
}
\arguments{
\item{seg_file}{segment file}

\item{gene_file}{gene to segment mapping}

\item{output_name}{name of output file will be postended with "_by_gene.txt"}
}
\description{
Reads file with features(PC's) in columns and samples in rows. And an annotation file.You pick a max # of features. Using caret
runs a prediction iteratively from 2 to max.features. Picks best predictor with least featurs. Returns samples correctly predicted
leaves out those samples incorrectly predicted.
}
