% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_ttest.R
\name{simple_ttest}
\alias{simple_ttest}
\title{does survival analysis}
\usage{
simple_ttest(dataset, fact, file.prefix)
}
\arguments{
\item{dataset}{data file, first column is features , columns samples, rows genes}

\item{fact}{first column sample names, second column type, must have no header}

\item{file.prefix}{name of output file will be postended with ".txt"}
}
\description{
Reads file with features(PC's) in columns and samples in rows. And an annotation file.You pick a max # of features. Using caret
runs a prediction iteratively from 2 to max.features. Picks best predictor with least featurs. Returns samples correctly predicted
leaves out those samples incorrectly predicted.
}
